% Generated by roxygen2 (4.0.1): do not edit by hand
\name{align2domain}
\alias{align2domain}
\title{align a sample to a model}
\usage{
align2domain(model, sample, scale = TRUE, ptDomain = NULL,
  ptSample = NULL)
}
\arguments{
\item{model}{statistical model of class "pPCA"}

\item{sample}{matrix or mesh3d}

\item{scale}{logical: request scaling during alignment}

\item{ptDomain}{integer vector: specifies the indices of the domain points that are to be used for registration (order is important).}

\item{ptSample}{integer vector: specifies the indices of the sample that are to be used  for registration (order is important).}
}
\value{
a rotated (and scaled) mesh or matrix - depending on the input.
}
\description{
align a sample to a model
}

