% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.r
\name{BSplineKernel}
\alias{BSplineKernel}
\title{create a scalar valued kernel of type BSplineKernel}
\usage{
BSplineKernel(support = 100, scale = 10)
}
\arguments{
\item{support}{support value for B-spline}

\item{scale}{scale factor}
}
\value{
object of class scalarKernel
}
\description{
create a scalar valued kernel of type BSplineKernel
}
\examples{
kernel1 <- BSplineKernel(100,50)
}
