% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.r
\name{CombineKernels}
\alias{CombineKernels}
\title{combine (add or multiply) two matrixKernels}
\usage{
CombineKernels(kernel1, kernel2, add = TRUE)
}
\arguments{
\item{kernel1}{object of class matrixKernel}

\item{kernel2}{object of class matrixKernel}

\item{add}{logical: if TRUE kernels will be added, multiplied otherwise}
}
\value{
object of class matrixKernel
}
\description{
combine (add or multiply) two matrixKernels
}
\examples{
isoKernel <- IsoKernel(DrawMean(mod),scale=0.001)
kernel1 <- MatrixValuedKernel(GaussianKernel(10),1)
combinedKernel <- CombineKernels(kernel1,isoKernel)
}

