% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/competingPoints.r
\docType{methods}
\name{competingPoints}
\alias{competingPoints}
\alias{competingPoints,pPCA,matrix,numeric-method}
\title{select the most probable points based on a statistical model}
\usage{
competingPoints(model, sample, index, mahalanobis = TRUE)

\S4method{competingPoints}{pPCA,matrix,numeric}(model, sample, index,
  mahalanobis = TRUE)
}
\arguments{
\item{model}{}

\item{sample}{k x 3 matrix of coordinates}

\item{index}{integer vector of lenght \code{k} assigning the corresponding index of the model's coordinates to each row of \code{sample}}

\item{mahlanobis}{logical: if FALSE, Euclideandistance is used.}
}
\value{
\item{mahadistance}{vector containing the mahalanobisdistances of all tested coordinates}
\item{goodverts}{the coordinates with the lowest mahalanobisdistance}
\item{goodrows}{integer vector containing the rows of \code{sample} that are selected}
\item{mahagood}{mahalanobisdistances of the probable coordinates only}
}
\description{
select the most probable points based on a statistical model, using the mahalanobisdistance
}
\examples{
require(Morpho)
     data(boneData)
    align <- rigidAlign(boneLM)$rotated
    mymod <- statismoBuildModel(align,representer=align[,,1],sigma=2,scale=TRUE)
#add some arbitrary data
myconfused <- matrix(rnorm(300),100,3)
myconfusedind <- sample(1:10,size=100,replace=T)
perturb <- sample(1:110)
out <- competingPoints(mymod,rbind(align[,,1],myconfused)[perturb,],c(1:10,myconfusedind)[perturb])
##check if the selected coords are identical to the actual ones
all.equal(align[,,1],out$goodverts)
}

