% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ComputeConstrainedModel}
\alias{ComputeConstrainedModel}
\title{Constrains a model of class pPCA by a subset of coordinates}
\usage{
ComputeConstrainedModel(x, model, align = FALSE, use.lm, deselect = FALSE,
  origSpace = FALSE)
}
\arguments{
\item{x}{a k x 3 matrix containing the coordinates of the reduces model}

\item{model}{an object of class \code{\link{pPCA}}}

\item{align}{logical: if TRUE, \code{x} will be aligned to the models mean}

\item{use.lm}{integer vector, specifying which coordinates from the full model are to be used/missing (see note)}

\item{deselect}{logical: if TRUE, use.lm specifies the missing coordinates instead of those present.}

\item{origSpace}{logical: if align=TRUE and origSpace=TRUE, the representer of the returned model will contain the estimated full shape in the original coordinate system of \code{x}}
}
\value{
an object of class pPCA constrained to \code{x}
}
\description{
Constrains a model of class pPCA by a subset of coordinates
}
\note{
if deselect = F, the order of the entries in \code{use.lm} is important: the i-th entry in use.lm specifies the index of the meanshapes coordinate belonging to the i-th coordinate of \code{x}.
}
\examples{
## create a model superimposed with missing landmarks 3 and 4
newmod <- pPCA(boneLM[,,-1],sigma=0,scale=TRUE,use.lm = 3:4,deselect=TRUE)
## predict the left out shape from the constrained model
boneLM1 <- ComputeConstrainedModel(boneLM[-c(3:4),,1],newmod,align=TRUE,use.lm=3:4,deselect=T,origSpace=TRUE)
## the coordinates of the estimated complete config are now stored in the representer's vertices
\dontrun{
##visualize prediction error
deformGrid3d(vert2points(boneLM1$representer),boneLM[,,1],ngrid=0)
}
}

