% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pPCA.r
\encoding{utf8}
\name{getCoordVar}
\alias{getCoordVar}
\title{get per coordinate variance from a statistical model}
\usage{
getCoordVar(model)
}
\arguments{
\item{model}{object of class pPCA}
}
\description{
get per coordinate variance from a statistical model
}
\note{
calculates the per-coordinate variance as described in Luethi(2009)
}
\references{
\enc{Lüthi}{Luethi} M, Albrecht T, Vetter T. 2009. Probabilistic modeling and visualization of the flexibility in morphable models. In: Mathematics of Surfaces XIII. Springer. p 251-264
}
