% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pPCA.r
\name{getDataLikelihood}
\alias{getCoefficients}
\alias{getDataLikelihood}
\alias{getDataLikelihood.matrix}
\alias{getDataLikelihood.mesh3d}
\title{calculate probability/coefficients for a matrix/mesh given a statistical model}
\usage{
getDataLikelihood(x, model, align = FALSE, lmDataset = NULL,
  lmModel = NULL)

\method{getDataLikelihood}{matrix}(x, model, align = FALSE,
  lmDataset = NULL, lmModel = NULL)

\method{getDataLikelihood}{mesh3d}(x, model, align = FALSE,
  lmDataset = NULL, lmModel = NULL)

getCoefficients(x, model, align = TRUE, lmDataset = NULL, lmModel = NULL)
}
\arguments{
\item{x}{matrix or mesh3d}

\item{model}{a model of class pPCA}

\item{align}{logical: if TRUE the data will be aligned to the model's mean}

\item{lmDataset}{optional: landmarks on the dataset used for alignment.}

\item{lmModel}{optional: landmarks on the model's mean used for alignment.}
}
\value{
\code{getDataLikelihood} returns a probability, while \code{getCoefficients} returns the (scaled) scores in the pPCA space.
}
\description{
calculate probability for a matrix/mesh given a statistical model
}
\details{
\code{getDataLikelihood} estimates the likelihood of a dataset for belonging to the model by exploiting the \eqn{\chi^2}{Chi-square}-distribution of the (squared) Mahalanobisdistance, which, in turn, is simply the squared norm of the sample's coefficients in the latent space.
}

