% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetFullJacobian.r
\name{GetFullJacobian}
\alias{GetFullJacobian}
\title{get Jacobian for all model points or a subset}
\usage{
GetFullJacobian(model, pt)
}
\arguments{
\item{model}{}

\item{pt}{integer vector specifying which points to use}
}
\value{
returns a matrix with rows n: (n+2) containing the Jacobian for the n-th point
}
\description{
get Jacobian for all model points or a subset
}
\seealso{
\code{\link{GetJacobian}}
}

