% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstrainedModel.r
\name{GetMahalanobisForPointSets}
\alias{GetMahalanobisForPointSets}
\title{Get pointwise Mahalanobis Distances based on Per-Coordinate Covariance Matrices}
\usage{
GetMahalanobisForPointSets(model, sample, pt)
}
\arguments{
\item{model}{shape model}

\item{sample}{matrix with sample landmarks to be tested}

\item{pt}{matrix with coordinates on the model mean or indices}
}
\value{
returns a vector of Mahalanobis distances
}
\description{
Get pointwise Mahalanobis Distances based on Per-Coordinate Covariance Matrices
}
\examples{
require(Morpho)
data(boneData)
align <- rigidAlign(boneLM)$rotated
mymod <- statismoBuildModel(align,representer=align[,,1],sigma=2,scale=TRUE)
mahadists <- GetMahalanobisForPointSets(mymod,align[,,1],1:10)
}
