% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAllClass.r
\docType{class}
\name{kernel-classes}
\alias{BSplineKernel-class}
\alias{CombinedKernel-class}
\alias{GaussianKernel-class}
\alias{IsoKernel-class}
\alias{MultiscaleBSplineKernel-class}
\alias{StatisticalModelKernel-class}
\alias{kernel-classes}
\title{Documentation of kernel classes}
\description{
Documentation of kernel classes
}
\details{
Blow are all S4 kernel classes and their containing slots, where the parameters are stored, that later used when passed to statismo. All kernel-classes have a common slot:
\itemize{
 \item{\code{kerneltype} - character: containing kernel specifics}
}

Available kernels are:
MultiscaleBSplineKernel-class
\itemize{
 \item{\code{support} - numeric: the BSpline support}
 \item{\code{scale} - numeric: scale factor}
 \item{\code{levels} - integer: Number of levels}
}

BSplineKernel-class:
\itemize{
 \item{\code{support} - numeric: the BSpline support}
}

GaussianKernel-class
\itemize{
 \item{\code{sigma} - numeric: kernel bandwidth}
 \item{\code{scale} - numeric: scale factor}
}

IsoKernel-class
\itemize{
 \item{\code{centroid} - vector: centroid around which to be scaled}
 \item{\code{scale} - numeric: scale factor}
}

CombinedKernel-class
\itemize{
 \item{\code{kernel} - list(s) containing the kernel above}
}
}

