% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.r
\name{MatrixValuedKernel}
\alias{MatrixValuedKernel}
\title{create a matrix valued kernel}
\usage{
MatrixValuedKernel(scalarKernel, scale = 1)
}
\arguments{
\item{scalarKernel}{object of class scalarKernel}

\item{scale}{scale factor of kernel}
}
\value{
object of class matrixKernel
}
\description{
create a matrix valued kernel
}
\examples{
gkernel <- GaussianKernel(2)
mvkernel <- MatrixValuedKernel(gkernel)
}

