% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mesh2vtp}
\alias{mesh2vtp}
\title{exports a triangular mesh of class mesh3d to a vtp file}
\usage{
mesh2vtp(mesh, filename = dataname)
}
\arguments{
\item{mesh}{mesh of class mesh3d}

\item{filename}{character}
}
\description{
exports a triangular mesh of class mesh3d to a vtp file
}

