% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meshalign}
\alias{meshalign}
\title{align meshes stored in a list by their vertices}
\usage{
meshalign(meshlist, scale = TRUE, tol = 1e-05)
}
\arguments{
\item{meshlist}{list containing triangular meshes of class "mesh3d"}

\item{scale}{logical: request scaling during alignment}
}
\description{
align meshes stored in a list by their vertices
}

