% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{modelinfo-class}
\alias{AddModelInfoParams}
\alias{AddModelInfoParams,modelinfo-method}
\alias{SetModelInfoParams<-}
\alias{modelinfo-class}
\title{Documentation of class modelinfo}
\usage{
AddModelInfoParams(x, value)

\S4method{AddModelInfoParams}{modelinfo}(x, value)

SetModelInfoParams(x) <- value

\S4method{SetModelInfoParams}{modelinfo}(x) <- value
}
\arguments{
\item{x}{an object of class "modelinfo"}

\item{value}{a list of or a single 2-valued character vectors}
}
\description{
Documentation of class modelinfo
}
\details{
The class has the following slots
\describe{
\item{datainfo}{a list containing 2-valued character vectors}
\item{parminfo}{a list containing 2-valued character vectors}
}
These can be modified using addParams and setParaminfo
with
}

