% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.r
\name{MultiscaleBSplineKernel}
\alias{MultiscaleBSplineKernel}
\title{create a scalar valued kernel of type MultiscaleBsplineKernel}
\usage{
MultiscaleBSplineKernel(support = 100, scale = 10, levels = 2)
}
\arguments{
\item{support}{suppor value for B-spline}

\item{scale}{scale factor}

\item{levels}{number of levels}
}
\value{
object of class scalarKernel
}
\description{
create a scalar valued kernel of type MultiscaleBSplineKernel
}
\examples{
kernel1 <- MultiscaleBSplineKernel(100,5)
}
