% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pPCA_pointer.r
\name{pPCA2pointer}
\alias{pPCA2pointer}
\alias{pointer2pPCA}
\title{convert an object of class pPCA to pPCA_pointer and vice versa}
\usage{
pPCA2pointer(x)

pointer2pPCA(x)
}
\arguments{
\item{x}{object of class pPCA or pPCA_pointer}
}
\value{
returns an object of class pPCA or pPCA_pointer
}
\description{
convert an object of class pPCA to pPCA_pointer and vice versa
}
