% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{pPCA-class}
\alias{pPCA-class}
\title{Documentation of class pPCA}
\description{
Documentation of class pPCA
}
\details{
The class contains the the follwing slots (still not yet set in stone)
\describe{
\item{PCA}{a list containing
\itemize{
\item{\code{sdev}: the square roots of the covariance matrix' eigenvalues}
\item{\code{rotation}: matrix containing the orthonormal PCBasis vectos}
\item{\code{x}: the scores within the latent space(scaled by 1/sdev)}
\item{\code{center}: a vector of the mean shape in with coordinates ordered

\code{(x1,y1,z1, x2, y2,z2, ..., xn,yn,zn)}}
 }
}
\item{scale}{logical: indicating if the data was aligned including scaling}
\item{representer}{an object of class mesh3d or a list with entry \code{vb} being a matrix with the columns containing coordinates and \code{it} a 0x0 matrix}
\item{sigma}{the noise estimation of the data}
\item{Variance}{a data.frame containing the Variance, cumulative Variance and Variance explained by each Principal component}
\item{rawdata}{optional data: a matrix with rows containing the mean centred coordinates in order \code{(x1,y1,z1, x2, y2,z2, ..., xn,yn,zn)}}
}
}

