% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAllClass.r
\docType{class}
\name{pPCA_pointer-class}
\alias{pPCA_pointer-class}
\title{Documentation of pPCA_pointer class}
\description{
Documentation of pPCA_pointer class
this class stores a pointer to an object of statismo::StatisticalModel<vtkPolyData>.
It has the following slots
\itemize{
 \item{\code{pointer} - externalptr: contains the external pointer}
 \item{\code{scale} - logical: contains information whether the data to build the model was aligned including scaling}
}
}
