% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{PredictSample}
\alias{PredictSample}
\alias{PredictSample,pPCA,matrix,ANY-method}
\alias{PredictSample,pPCA,mesh3d,logical-method}
\title{predict or restrict a mesh or matrix based on a statistical model}
\usage{
PredictSample(model, dataset, representer = TRUE, ...)

\S4method{PredictSample}{pPCA,matrix,ANY}(model, dataset, representer = TRUE,
  origSpace = TRUE, use.lm = NULL, deselect = FALSE, sdmax,
  mahaprob = c("none", "chisq", "dist"), align = TRUE, ...)

\S4method{PredictSample}{pPCA,mesh3d,logical}(model, dataset,
  representer = TRUE, origSpace = TRUE, use.lm = NULL, deselect = FALSE,
  sdmax, mahaprob = c("none", "chisq", "dist"), align = TRUE, ...)
}
\arguments{
\item{model}{model of class \code{pPCA}}

\item{dataset}{a matrix or a mesh3d}

\item{representer}{if TRUE and the model contains a representer mesh, a surface mesh will be returned, coordinate matrix otherwise.}

\item{origSpace}{logical: rotate the estimation back into the original coordinate system.}

\item{pPCA}{logical: if TRUE, a constrained pPCA model is returned.
"chisq" uses the Chi-Square distribution of the squared Mahalanobisdistance, while "dist" restricts the values to
be within a multi-dimensional sphere of radius \code{sdmax}. If FALSE the probability will be determined per PC separately.}

\item{use.lm}{optional: integer vector specifying row indices of the coordinates to use for rigid registration on the model's meanshape.}

\item{sdmax}{maximum allowed standard deviation (per Principal axis) within the model space. Defines the probabilistic boundaries.}

\item{mahaprob}{character: if != "none", use mahalanobis-distance to determine overall probability (of the shape projected into the model space.}
}
\value{
\code{PredictSample} returns a matrix/mesh3d restricted to the boundaries given by the modelspace.
}
\description{
predict or restrict a mesh or matrix based on a statistical model
}
\seealso{
\code{\link{StatismoModelMembers}}
}

