% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.r
\name{ProductKernels}
\alias{ProductKernels}
\title{Multiply two kernels}
\usage{
ProductKernels(kernel1, kernel2)
}
\arguments{
\item{kernel1}{object of class matrixKernel}

\item{kernel2}{object of class matrixKernel}
}
\value{
object of class CombinedKernel
}
\description{
Multiply two kernels
}

