% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePointsFromModel.r
\name{removePointsFromModel}
\alias{removePointsFromModel}
\title{remove a set of point from a statismo shape model}
\usage{
removePointsFromModel(model, pointind)
}
\arguments{
\item{model}{pPCA model}

\item{pointind}{integer vector indicating which coordinates to remove}
}
\value{
returns and updated model
}
\description{
remove a set of point from a statismo shape model
}
\examples{
## create a GP model from a mesh and remove the first 1000 coordinates
require(Rvcg)
data(humface)
hummodel <- statismoModelFromRepresenter(humface,kernel=list(c(50,50)))
hummodel0 <- removePointsFromModel(hummodel,1:1000)
\dontrun{
require(rgl)
shade3d(DrawSample(hummodel0),col=3)
}
}

