% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pPCAhelper.r
\docType{methods}
\name{representer2sample}
\alias{representer2sample}
\alias{representer2sample,pPCA-method}
\title{get the representer from a model of class "pPCA"}
\usage{
representer2sample(model)

\S4method{representer2sample}{pPCA}(model)
}
\arguments{
\item{model}{object of class \code{\link{pPCA}}}
}
\value{
an object of class mesh3d or matrix, depending whether a point cloud or a triangular mesh is the model's representer.
}
\description{
get the representer from a model of class "pPCA"
}

