% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/restoreSamples.r
\name{restoreSamples}
\alias{restoreSamples}
\title{restores original samples from a model with stored PC-scores}
\usage{
restoreSamples(model)
}
\arguments{
\item{model}{shape model model of class "pPCA"}
}
\value{
returns a list of restored shapes
}
\description{
restores original samples (meshes or pointclouds) from a model with stored PC-scores
}

