% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rigidAlign}
\alias{rigidAlign}
\title{Fast Procrustes align of coordinates}
\usage{
rigidAlign(array, scale = TRUE, missingIndex, deselect = FALSE)
}
\arguments{
\item{array}{array of coordinates}

\item{scale}{logical: request scaling during alignment}

\item{missingIndex}{integer vector: specifies which points are missing (for building constrained model)}

\item{deselect}{logical: if TRUE, missingIndex references the existing coordinates instead of the missing ones.}
}
\value{
a list containing
\item{rotated}{array containing registered coordinates}
\item{mshape}{matrix containing meanshape}
}
\description{
Fast Procrustes align of coordinates
}

