\docType{package}
\name{Rvtk-package}
\alias{Rvtk}
\alias{Rvtk-package}
\title{Interface between R and vtk-libraries for mesh operations}
\description{
Aims to provide meshing functionality from vtk for R.
}
\details{
\tabular{ll}{ Package: \tab Rvtk\cr Type: \tab Package\cr
Version: \tab 0.1\cr Date: \tab 2014-03-25\cr License: \tab
GPL\cr LazyLoad: \tab yes\cr }
}
\author{
Stefan Schlager

Maintainer: Stefan Schlager <zarquon42@gmail.com>
}
\references{
To be announced
}
\keyword{package}

