% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvtkStatismo-package.R
\name{deprecated}
\alias{deprecated}
\alias{ComputeCoefficientsForDataset}
\alias{ComputeLogProbabilityOfDataset}
\alias{ComputeMahalanobisDistanceForDataset}
\alias{read.fcsv}
\alias{write.fcsv}
\title{deprecated functions of Morpho}
\usage{
ComputeCoefficientsForDataset(...)

ComputeLogProbabilityOfDataset(...)

ComputeMahalanobisDistanceForDataset(...)

read.fcsv(...)

write.fcsv(...)
}
\description{
document deprecated functions
}
\keyword{internal}
