% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{statismoConstrainModel}
\alias{statismoConstrainModel}
\title{constrain a model of class pPCA}
\usage{
statismoConstrainModel(model, sample, pt, ptValueNoise)

\S4method{statismoConstrainModel}{pPCA,matrix,matrix}(model, sample, pt,
  ptValueNoise)

\S4method{statismoConstrainModel}{pPCA,matrix,numeric}(model, sample, pt,
  ptValueNoise)

\S4method{statismoConstrainModel}{pPCA,numeric,numeric}(model, sample, pt,
  ptValueNoise)
}
\arguments{
\item{model}{object of class pPCA}

\item{sample}{k x 3 matrix containing coordinates to constrain model to}

\item{pt}{either a matrix with each row containing points on the model's domain corresponding to the row in \code{sample} or an integer vector specifying the coordinates of the sample's mean corresponding to \code{sample}}

\item{ptValueNoise}{numeric: specify noise on constraints.}
}
\value{
a constrained model
}
\description{
constrain a model of class pPCA
}

