% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstrainedModel.r, R/ConstrainedModelPointer.r
\docType{methods}
\name{statismoConstrainModelSafe}
\alias{statismoConstrainModelSafe}
\alias{statismoConstrainModelSafe,pPCA,matrix,matrix-method}
\alias{statismoConstrainModelSafe,pPCA,matrix,numeric-method}
\alias{statismoConstrainModelSafe,pPCA,numeric,numeric-method}
\alias{statismoConstrainModelSafe,pPCA_pointer,matrix,matrix-method}
\alias{statismoConstrainModelSafe,pPCA_pointer,matrix,numeric-method}
\alias{statismoConstrainModelSafe,pPCA_pointer,numeric,numeric-method}
\title{calculate a posterior model but only use likely correspondences}
\usage{
statismoConstrainModelSafe(model, sample, pt, ptValueNoise, sdmax = 5,
  pointer = FALSE)

\S4method{statismoConstrainModelSafe}{pPCA,matrix,numeric}(model, sample, pt,
  ptValueNoise, sdmax = 5, pointer = FALSE)

\S4method{statismoConstrainModelSafe}{pPCA,matrix,matrix}(model, sample, pt,
  ptValueNoise, sdmax = 5, pointer = FALSE)

\S4method{statismoConstrainModelSafe}{pPCA,numeric,numeric}(model, sample, pt,
  ptValueNoise, sdmax = 5, pointer = FALSE)

\S4method{statismoConstrainModelSafe}{pPCA_pointer,matrix,numeric}(model,
  sample, pt, ptValueNoise, sdmax = 5, pointer = TRUE)

\S4method{statismoConstrainModelSafe}{pPCA_pointer,matrix,matrix}(model, sample,
  pt, ptValueNoise, sdmax = 5, pointer = TRUE)

\S4method{statismoConstrainModelSafe}{pPCA_pointer,numeric,numeric}(model,
  sample, pt, ptValueNoise, sdmax = 5, pointer = TRUE)
}
\arguments{
\item{model}{object of class pPCA}

\item{sample}{matrix containing coordinates to constrain model to}

\item{pt}{either a k x 3 matrix with each row containing points on the model's domain corresponding to the row in \code{sample} or an integer vector specifying the coordinates of the sample's mean corresponding to \code{sample}}

\item{ptValueNoise}{numeric: specify noise on constraints.}

\item{sdmax}{a measure in standard deviations to allow the likelihood of the correspondeces between sample and model. (using chi-square distribution)}
}
\value{
a constrained model
}
\description{
calculate a posterior model but only use likely correspondences
}

