% Generated by roxygen2 (4.0.1): do not edit by hand
\name{statismoGPmodel}
\alias{statismoGPmodel}
\title{expands a models variability by adding a Gaussian kernel function}
\usage{
statismoGPmodel(model, kernel = list(c(100, 70)), ncomp = 10,
  nystroem = 500)
}
\arguments{
\item{model}{shape model of class "pPCA"}

\item{kernel}{a list containing two valued vectors containing with the first entry specifiying the bandwidth and the second the scaling of the Gaussian kernels (currently only the first list entry is used)}

\item{ncomp}{integer: number of PCs to approximate}

\item{nystroem}{number of samples to compute Nystroem approximation of eigenvectors}
}
\value{
returns a shape model of class "pPCA"
require(Morpho)
data(boneData)
mod <- pPCA(boneLM)
GPmod <- statismoGPmodel(mod)##extend flexibility
}
\description{
expands a models variability by adding a Gaussian kernel function to the empiric covariance matrix and builds a low-rank approximation of the resulting PCA
}

