% Generated by roxygen2 (4.0.1): do not edit by hand
\name{StatismoMatrices}
\alias{GetCovarianceMatrix}
\alias{GetJacobian}
\alias{GetOrthonormalPCABasisMatrix}
\alias{GetProjectionMatrix}
\alias{StatismoMatrices}
\title{Get Matrices from StatisticalModel class}
\usage{
GetPCABasisMatrix(model)

GetOrthonormalPCABasisMatrix(model)

GetCovarianceMatrix(model)

GetJacobian(model, pt)

GetProjectionMatrix(model)
}
\arguments{
\item{model}{model of class "pPCA"}

\item{pt}{either an integer pointing to the index of the domain or a numeric vector of length 3 specifying a point on the domain of the model}

\item{pt1}{either an integer pointing to the index of the domain or a numeric vector of length 3 specifying a point on the domain of the model}

\item{pt2}{either an integer pointing to the index of the domain or a numeric vector of length 3 specifying a point on the domain of the model}
}
\value{
\item{GetPCABasisMatrix}{returns the (scaled) Basis of the latent space}
\item{GetOrthonormalPCABasisMatrix}{ returns the orthonormal Basis of the latent space}
 \item{GetCovarianceAtPoint}{returns the 3 x 3 covariance matrix for \code{pt1} and \code{pt2}}
\item{GetJacobian}{ returns the 3 x 3 Jacobian matrix at \code{pt}}
\item{GetProjectionMatrix}{ returns matrix to project a sample vector into the latent space (this is not a member function but might prove useful anyway)}
}
\description{
Get Matrices from StatisticalModel class - such as projection matrices, covariance matrices or Jacobian
}

