% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statismoModelFromRepresenter.r
\name{statismoModelFromRepresenter}
\alias{statismoModelFromRepresenter}
\title{generate model from a representer using gaussian kernels}
\usage{
statismoModelFromRepresenter(representer, kernel = list(c(100, 70)),
  ncomp = 10, nystroem = 500, combine = "sum", isoScale = 0,
  centroid = NULL)
}
\arguments{
\item{representer}{mesh3d or matrix used as representer}

\item{kernel}{a list containing 2 numeric vectors containing. Except the first entry may be 3 valued and then interpreted as Multiscale Bspline kernel. For a Gaussian Kernel, the first entry specifies the bandwidth and the second the scaling. For a Multiscale kernel, the additional 3rd entry sets the number of levels.}

\item{ncomp}{integer: number of PCs to approximate}

\item{nystroem}{number of samples to compute Nystroem approximation of eigenvectors}

\item{combine}{character determining how to combine the kernels: "sum" or "product" are supported.}

\item{isoScale}{standard deviation of isotropic scaling.}

\item{centroid}{specify the center of scaling. If NULL, the centroid will be used.}
}
\value{
returns a shape model of class \code{\link{pPCA}}
}
\description{
generate model from a representer using gaussian kernels
}
\examples{
require(Rvcg)
data(humface)
hummodel <- statismoModelFromRepresenter(humface)
\dontrun{
require(rgl)
for (i in 1:5) wire3d(DrawSample(hummodel),col=i)
}
}

