% Generated by roxygen2 (4.0.1): do not edit by hand
\name{StatismoParameters}
\alias{GetMeanVector}
\alias{GetPCAVarianceVector}
\alias{StatismoParameters}
\title{Get model parameters}
\usage{
GetNoiseVariance(model)

GetMeanVector(model)

GetPCAVarianceVector(model)
}
\arguments{
\item{model}{model of class "pPCA"}
}
\value{
\item{GetNoiseVariance}{returns the estimated noise in the model}
\item{GetPCAVarianceVector}{returns the variance in the model}
\item{GetMeanVector}{returns the model's mean vector}
}
\description{
Get model parameters such as variance or noise variance
}

