% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r, R/ModelMembers.r
\docType{methods}
\name{StatismoSample}
\alias{ComputeLogProbability}
\alias{ComputeLogProbability,pPCA-method}
\alias{ComputeMahalanobisDistance}
\alias{ComputeMahalanobisDistance,pPCA-method}
\alias{ComputeProbability}
\alias{ComputeProbability,pPCA-method}
\alias{StatismoSample}
\title{Retrieve information about a sample from the model}
\usage{
ComputeLogProbability(model, dataset)

ComputeProbability(model, dataset)

ComputeMahalanobisDistance(model, dataset)

\S4method{ComputeLogProbability}{pPCA}(model, dataset)

\S4method{ComputeProbability}{pPCA}(model, dataset)

\S4method{ComputeMahalanobisDistance}{pPCA}(model, dataset)
}
\arguments{
\item{model}{model of class "pPCA"}

\item{dataset}{a matrix or mesh3d aligned to the model's mean}
}
\value{
\item{ComputeLogProbability}{returns the log-probability density for the sample}
\item{ComputeProbability}{returns the probability density for the sample}
}
\description{
Retrieve information about a sample from the model
}
\seealso{
\code{\link{getDataLikelihood}}
}

