% Generated by roxygen2 (4.0.1): do not edit by hand
\name{statismoSaveModel}
\alias{statismoSaveModel}
\title{save and load a statistical model of class pPCA to statismo hdf5 format}
\usage{
statismoSaveModel(model, modelname)
}
\arguments{
\item{model}{object of class pPCA}

\item{modelname}{filename to read/save}
}
\value{
statismoLoadModel returns an object of class "pPCA"
}
\description{
save and load a statistical model of class pPCA to statismo hdf5 format
}

