% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.r
\name{StatisticalModelKernel}
\alias{StatisticalModelKernel}
\title{Create empirical StatisticalModelKernel}
\usage{
StatisticalModelKernel()
}
\value{
returns object of class StatisticalModelKernel
}
\description{
Create empirical StatisticalModelKernel
}
