% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.r
\name{SumKernels}
\alias{SumKernels}
\title{Add two kernels}
\usage{
SumKernels(kernel1, kernel2)
}
\arguments{
\item{kernel1}{object of class matrixKernel}

\item{kernel2}{object of class matrixKernel}
}
\value{
object of class CombinedKernel
}
\description{
Add two kernels
}

