% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkBooleanOp.r
\name{vtkBooleanOp}
\alias{vtkBooleanOp}
\title{run boolean operations on mesh}
\usage{
vtkBooleanOp(mesh1, mesh2, type = 0, reorient = TRUE, tol = 1e-06)
}
\arguments{
\item{mesh1}{triangular mesh}

\item{mesh2}{triangular mesh}

\item{type}{integer: boolean operation. 0=Union, 1=intersection, 2=difference}

\item{reorient}{logical: Turn on/off cell reorientation of the intersection portion of the surface when the type=2 (DIFFERENCE). Defaults to TRUE.}

\item{tol}{Set the tolerance used to determine when a point's absolute distance is considered to be zero. Defaults to 1e-6.}
}
\value{
returns manipulated mesh
}
\description{
run boolean operations on mesh
}
\examples{
require(Rvcg);require(Morpho)
sphere <- vcgSphere()
myplane <- list(vb=rbind(cbind(c(1,0,0),c(0,1,0),c(0,0,1)),1),
                it=as.matrix(c(1:3)));class(myplane) <- "mesh3d"
myplane <- scalemesh(myplane,5)
cutsphere <- vtkBooleanOp(sphere,myplane,type = 2)
}

