% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkExtractOutsideMesh.r
\name{vtkExtractOutsideMesh}
\alias{vtkExtractOutsideMesh}
\title{get outer mesh}
\usage{
vtkExtractOutsideMesh(x)
}
\arguments{
\item{x}{mesh}
}
\value{
returns only those segments visible from the outside of the mesh
}
\description{
get outer mesh and discard enclosed segments
}

