% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vtkFillHoles}
\alias{vtkFillHoles}
\title{fill holes in triangular mesh}
\usage{
vtkFillHoles(mesh, holesize = 1e+06)
}
\arguments{
\item{mesh}{triangular mesh}

\item{holesize}{maximal size of holes to fill}
}
\value{
mesh with filled holes
}
\description{
fill holes in triangular mesh
}
\details{
see documentation of vtkFillHolesFilter for details
}

