% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkGeodesicPath.r
\name{vtkGeodesicPath}
\alias{vtkGeodesicPath}
\title{Compute Geodesic path between two vertices}
\usage{
vtkGeodesicPath(x, start, end)
}
\arguments{
\item{x}{triangular mesh of class mesh3d}

\item{start}{index of vertex to start}

\item{end}{index of vertex to end}
}
\value{
list containing
\item{distance}{geodesic distance (cummulative distance between all vertices along the way)}
\item{index}{vector containing vertex indices from start to end}
}
\description{
Compute Geodesic path between two vertices
}
\examples{
require(Rvcg)
data(humface)
gp <- vtkGeodesicPath(humface,1,1000)
\dontrun{
require(rgl);require(Morpho)
points3d(vert2points(humface)[gp$index,])
lines3d(vert2points(humface)[gp$index,],col="blue",lwd=2)
spheres3d(vert2points(humface)[c(1,1000),],col=2)
wire3d(humface,col=3)
}
}
