% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkGeodesicPath.r
\name{vtkGeodesicPathForPointPair}
\alias{vtkGeodesicPathForPointPair}
\title{Compute Geodesic path between two points on a mesh based on the closest vertices}
\usage{
vtkGeodesicPathForPointPair(x, start, end)
}
\arguments{
\item{x}{triangular mesh of class mesh3d}

\item{start}{vector of length 3 containing starting position}

\item{end}{vector of length 3 containing starting position}
}
\value{
list containing
\item{distance}{geodesic distance (cummulative distance between all vertices along the way)}
\item{index}{vector containing vertex indices from start to end}
}
\description{
Compute Geodesic path between two points on a mesh based on the closest vertices
}
\examples{
require(Rvcg)
data(humface)
gp <- vtkGeodesicPathForPointPair(humface,humface.lm[1,],humface.lm[2,])
\dontrun{
require(rgl)
points3d(vert2points(humface)[gp$index,])
lines3d(vert2points(humface)[gp$index,],col="blue",lwd=2)
spheres3d(humface.lm[1:2,],col=2)
wire3d(humface,col=3)
}
}

