% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtkICP.r
\name{vtkICP}
\alias{vtkICP}
\title{Iteratively closest point registration}
\usage{
vtkICP(refmesh, tarmesh, iterations = 10, center = FALSE, type = c("none",
  "scale", "affine"), sample = 200)
}
\arguments{
\item{refmesh}{moving mesh}

\item{tarmesh}{fixed mesh}

\item{center}{logical: rotate meshes around center of gravity.}

\item{type}{type of transform to be used: "none"=rigid,"scale"=similarity transform, "affine"=affine transform.}

\item{sample}{number of random samples to use}

\item{iteration}{integer: number of iterations to run}
}
\description{
Iteratively closest point registration
}

