% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtkImageBlend.r
\name{vtkImageBlend}
\alias{vtkImageBlend}
\title{Blend two Images using vtkImageBlend}
\usage{
vtkImageBlend(image1, image2, outname = "out.mha")
}
\arguments{
\item{image1}{character: path to first image}

\item{image2}{character: path to second image}

\item{outname}{character: image to save merged image to.}
}
\description{
Blend two Images using vtkImageBlend
}

