% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vtkImageTransform}
\alias{vtkImageTransform}
\title{Transform an image based on landmark configuration}
\usage{
vtkImageTransform(image, sourceLM, targetLM, outname = "default.mha",
  type = c("none", "scale", "affine"), RAS2IJK = diag(c(-1, -1, 1)),
  interpolation = 2)
}
\arguments{
\item{image}{path to 3D-image (tiff, mhd)}

\item{sourceLM}{the landmarks on the image}

\item{targetLM}{the target landmarks}

\item{outname}{name of the file to be saved (must end with .mha)}

\item{type}{type of transformation}

\item{RAS2IJK}{3x3 transformation matrix from point to image space}

\item{interpolation}{type of image interpolation 0=Nearest Neighbour, 1=Linear , 2=Cubic}
}
\value{
returns (invisible)  TRUE if successful, else FALSE
}
\description{
Transform an image based on landmark configuration
}

