% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtkstuff.r
\name{read.vtk}
\alias{read.vtk}
\title{imports vtk, vtp and wrl files containing meshes}
\usage{
read.vtk(filename)
}
\arguments{
\item{filename}{character string}
}
\value{
triangular mesh of class mesh3d
}
\description{
imports vtk, vtp and wrl files containing meshes
}
\note{
read.wrl is simply an alias for read.vtk
}

