% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtkMesh2Image.r
\name{vtkMesh2Image}
\alias{vtkMesh2Image}
\title{convert a 3D triangular mesh into a voxel image}
\usage{
vtkMesh2Image(mesh, spacing = c(0.5, 0.5, 0.5), filename = "default.mha",
  IJK2RAS = diag(c(-1, -1, 1)), margin = 0.1, col = 255)
}
\arguments{
\item{mesh}{triangular mesh of class mesh3d}

\item{spacing}{numeric vector of length 3: voxel spacings}

\item{filename}{filename to save as. Can be mha/mhd for VTK_VERSION < 6.2 and
additionally nii or nii.gz else}

\item{IJK2RAS}{3x3 transformation matrix from point to image space}

\item{margin}{percentage of margin to add around the volume}

\item{col}{integer: value 0 <= col <= 255. Defines the foreground color.}
}
\description{
convert a 3D triangular mesh into a voxel image
}
\examples{
require(Rvcg)
data(dummyhead)
vtkMesh2Image(dummyhead.mesh)
}

