% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtkSegment2PolyData.r
\name{vtkSegment}
\alias{vtkSegment}
\title{create Isosurface from image}
\usage{
vtkSegment(file, value = 1, IJK2RAS = diag(c(-1, -1, 1, 1)))
}
\arguments{
\item{file}{3D-image file mha/nii.gz - depending of your system's VTK version}

\item{value}{isovalue}

\item{IJK2RAS}{4x4 IJK2RAS transform}
}
\value{
returns a triangular mesh of class mesh3d
}
\description{
create Isosurface from image
}

