% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkExtractOutsideMesh.r
\name{vtkSurfaceReko}
\alias{vtkSurfaceReko}
\title{create surface from pointcloud}
\usage{
vtkSurfaceReko(x, sampleSpacing = 0)
}
\arguments{
\item{x}{pointcloud or mesh}

\item{sampleSpacing}{set spacing for contour creation (0=auto)}
}
\value{
mesh of class mesh3d
}
\description{
create surface from pointcloud or mesh
}

