% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkTriangulate.r
\name{vtkTriangulate}
\alias{vtkTriangulate}
\title{create Isosurface from image}
\usage{
vtkTriangulate(file, value = 1, IJK2RAS = diag(c(-1, -1, 1, 1)))
}
\arguments{
\item{file}{3D-image file mha/nii.gz - depending of your system's VTK version}

\item{value}{isovalue}

\item{IJK2RAS}{4x4 IJK2RAS transform}
}
\value{
returns a triangular mesh of class mesh3d
}
\description{
create Isosurface from image
}

