% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDataLikelihood}
\alias{getCoefficients}
\alias{getDataLikelihood}
\alias{getDataLikelihood.matrix}
\alias{getDataLikelihood.mesh3d}
\title{calculate probability/coefficients for a matrix/mesh given a statistical model}
\usage{
getDataLikelihood(x, model, align = FALSE, use.lm)

\method{getDataLikelihood}{matrix}(x, model, align = FALSE, use.lm = NULL)

\method{getDataLikelihood}{mesh3d}(x, model, align = FALSE, use.lm = NULL)

getCoefficients(x, model, align = TRUE, use.lm = NULL)
}
\arguments{
\item{x}{matrix or mesh3d}

\item{model}{a model of class pPCA}

\item{align}{logical: if TRUE the data will be aligned to the model's mean}

\item{use.lm}{integer vector specifying row indices of the coordinates to use for rigid registration on the model's meanshape.}
}
\value{
\code{getProb} returns a probability, while \code{getCoefficients} returns the (scaled) scores in the pPCA space.
}
\description{
calculate probability for a matrix/mesh given a statistical model
}

