% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.r
\name{IsoKernel}
\alias{IsoKernel}
\title{create an isotropic kernel}
\usage{
IsoKernel(x, scale = 0.01)
}
\arguments{
\item{x}{matrix or mesh based on which the scaling is centered}

\item{scale}{scale factor}
}
\value{
object of class matrixKernel
require(Rvcg)
data(humface)
isokernel <- IsoKernel(humface,scale=0.01)
}
\description{
create an isotropic kernel
}

