% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.r
\name{IsoKernel}
\alias{IsoKernel}
\title{create an isotropic kernel}
\usage{
IsoKernel(scale = 0.01, x = NULL, centroid = NULL)
}
\arguments{
\item{scale}{scale factor}

\item{x}{matrix or mesh to calculate the centroid from (overrides centroid)}

\item{centroid}{centroid}
}
\value{
object of class matrixKernel
require(Rvcg)
data(humface)
isokernel <- IsoKernel(humface,scale=0.01)
}
\description{
create an isotropic kernel
}

