% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statismoBuildConditionalModel.r
\name{manageConditioningData}
\alias{manageConditioningData}
\title{convert a data.frame or vector of factors into a format usable by statismoBuildConditionalModel}
\usage{
manageConditioningData(trainingData)
}
\arguments{
\item{trainingData}{training data, can be data.frame, character vector/matrix, etc.}
}
\value{
\item{trainingData}{Data converted into numeric matrix}
\item{surrogateInfo}{vector of 0 and 1 indicating which input was categorial/continuous}
\item{encode}{in case categorial variables where encoded into integers, this lists the association of the categorial values to the corresponding dummy encoding.}
\item{order}{order of variables after resorting}
}
\description{
convert a data.frame or vector of factors into a format usable by statismoBuildConditionalModel and extract variable information.
}
\note{
the variables will be resorted with the first being continuous variables and then categorial ones.
}
\seealso{
\code{\link{statismoBuildConditionalModel}}
}
