% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meshalign}
\alias{meshalign}
\title{align meshes stored in a list by their vertices}
\usage{
meshalign(meshlist, scale = TRUE, array = FALSE)
}
\arguments{
\item{meshlist}{list containing triangular meshes of class "mesh3d"}

\item{scale}{logical: request scaling during alignment}

\item{array}{logical: if TRUE the superimposed vertices will be returned as 3D array.}
}
\value{
returns a list of aligned meshes or an array of dimensions k x 3 x n, where k=number of vertices and n=sample size.
}
\description{
align meshes stored in a list by their vertices
}

