% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alignment.r
\name{meshalign}
\alias{meshalign}
\title{align meshes stored in a list by their vertices}
\usage{
meshalign(meshlist, scale = FALSE, use.lm = NULL, deselect = FALSE,
  array = FALSE)
}
\arguments{
\item{meshlist}{list containing triangular meshes of class "mesh3d"}

\item{scale}{logical: request scaling during alignment}

\item{use.lm}{integer vector: specifies the indices of the points that are to be used in the constrained model}

\item{deselect}{logical: if TRUE, missingIndex references the existing coordinates instead of the missing ones.}

\item{array}{logical: if TRUE the superimposed vertices will be returned as 3D array.}
}
\value{
returns a list of aligned meshes or an array of dimensions k x 3 x n, where k=number of vertices and n=sample size.
}
\description{
align meshes stored in a list by their vertices
}

