% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alignment.r
\name{meshalign}
\alias{meshalign}
\title{align meshes stored in a list by their vertices}
\usage{
meshalign(meshlist, scale = FALSE, use.lm = NULL, deselect = FALSE,
  array = FALSE, reference = NULL, reflection = FALSE)
}
\arguments{
\item{meshlist}{list containing triangular meshes of class "mesh3d"}

\item{scale}{logical: request scaling during alignment}

\item{use.lm}{integer vector: specifies the indices of the points that are to be used in the constrained model}

\item{deselect}{logical: if TRUE, missingIndex references the existing coordinates instead of the missing ones.}

\item{array}{logical: if TRUE the superimposed vertices will be returned as 3D array.}

\item{reference}{integer: instead of a GPA, all specimens are aligned to this specific one.}

\item{reflection}{logical: if TRUE, reflections are allowed to minimzie squared distance.}
}
\value{
returns a list of aligned meshes or an array of dimensions k x 3 x n, where k=number of vertices and n=sample size.
}
\description{
align meshes stored in a list by their vertices
}
\examples{
mod <- statismoLoadModel(system.file("extdata","mandibles.h5",package="RvtkStatismo"))
shapes <- restoreSamples(mod)
newalign <- meshalign(shapes,reference=1)
}

