% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/statismoHelper.r
\name{meshlist2array}
\alias{meshlist2array}
\title{convert meshes to array consisting of vertex coordinates}
\usage{
meshlist2array(meshlist)
}
\arguments{
\item{meshlist}{list containing triangular meshes of class "mesh3d"}
}
\value{
returns an array with k x 3 x n dimensions where k=number of vertices, and n=sample size.
}
\description{
convert meshes to array consisting of vertex coordinates
}

