% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{modelinfo-class}
\alias{modelinfo-class}
\title{Documentation of class modelinfo}
\description{
Documentation of class modelinfo
\describe{
\item{\code{datainfo}: a list containing 2-valued character vectors}
\item{\code{paraminfo}: a list containing 2-valued character vectors}
}
}

