% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.r
\name{MultiscaleBsplineKernel}
\alias{MultiscaleBsplineKernel}
\title{create a scalar valued kernel of type MultiscaleBsplineKernel}
\usage{
MultiscaleBsplineKernel(support = 100, levels = 2)
}
\arguments{
\item{support}{suppor value for B-spline}

\item{levels}{number of levels}
}
\value{
object of class scalarKernel
}
\description{
create a scalar valued kernel of type MultiscaleBsplineKernel
}
\examples{
kernel1 <- MultiscaleBsplineKernel(100,5)
}

