% Generated by roxygen2 (4.0.1): do not edit by hand
\encoding{utf8}
\name{pPCA}
\alias{pPCA}
\alias{setMod}
\alias{setMod.pPCA}
\title{calculate or modify a probablistic PCA based on 3D-coordinates}
\usage{
pPCA(array, align = TRUE, use.lm = NULL, deselect = FALSE, sigma = NULL,
  exVar = 1, scale = TRUE, representer = NULL)

setMod(procMod, sigma, exVar)

\method{setMod}{pPCA}(model, sigma = NULL, exVar = 1)
}
\arguments{
\item{array}{array of dimensions k x 3 x n, where k=number of coordinates and n=sample size.}

\item{align}{logical: if TRUE, the data will be aligned first}

\item{use.lm}{integer vector: specifies the indices of the points that are to be used in the constrained model}

\item{deselect}{logical: if TRUE, \code{use.lm} references the missing coordinates instead of the present ones.}

\item{sigma}{estimate of error variance (sensible is a value estimating coordinate error in terms of observer error)}

\item{exVar}{numeric value with \code{0 < exVar <= 1} specifying the PCs to be included by their cumulative explained Variance}

\item{scale}{logical: allow scaling in Procrustes fitting}

\item{fullfit}{logical: if FALSE only the non-missing points will be used for registration.}

\item{representer}{a triangular mesh, where the vertices correspond to the coordinates in \code{array}, leave NULL for pointclouds.}

\item{model}{object of class \code{pPCA}}
}
\value{
returns a probabilistic PCA model of class "pPCA".
\code{setMod} is used to modify existing models by changing sigma and exVar.


The class \code{"pPCA"} is a list containing the follwing items (still not yet set in stone)
\item{PCA}{a list containing
\itemize{
\item{\code{sdev}: the square roots of the covariance matrix' eigenvalues}
\item{\code{rotation}: matrix containing the orthonormal PCBasis vectos}
\item{\code{x}: the scores within the latent space(scaled by 1/sdev)}
\item{\code{center}: a vector of the mean shape in with coordinates ordered

\code{(x1,y1,z1, x2, y2,z2, ..., xn,yn,zn)}}
 }
}
\item{scale}{logical: indicating if the data was aligned including scaling}
\item{representer}{an object of class mesh3d or a list with entry \code{vb} being a matrix with the columns containing coordinates and \code{it} a 0x0 matrix}
\item{sigma}{the noise estimation of the data}
\item{Variance}{a data.frame containing the Variance, cumulative Variance and Variance explained by each Principal component}
\item{rawdata}{optional data: a matrix with rows containing the mean centred coordinates in order \code{(x1,y1,z1, x2, y2,z2, ..., xn,yn,zn)}}
}
\description{
calculate or modify a probablistic PCA based on 3D-coordinates
}
\examples{
require(Morpho)
data(boneData)
model <- pPCA(boneLM[,,])
## change parameters without recomputing Procrustes fit
model1 <- setMod(model, sigma=1, exVar=0.8)
}
\references{
\enc{Lüthi}{Luethi} M, Albrecht T, Vetter T. 2009. Probabilistic modeling and visualization of the flexibility in morphable models. In: Mathematics of Surfaces XIII. Springer. p 251-264
}

