% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\encoding{utf8}
\name{pPCA}
\alias{UpdateModel}
\alias{UpdateModel,pPCA-method}
\alias{pPCA}
\title{calculate or modify a probablistic PCA based on 3D-coordinates}
\usage{
pPCA(array, align = TRUE, use.lm = NULL, deselect = FALSE, sigma = NULL,
  exVar = 1, scale = FALSE, representer = NULL)

UpdateModel(model, sigma = NULL, exVar = 1)

\S4method{UpdateModel}{pPCA}(model, sigma = NULL, exVar = 1)
}
\arguments{
\item{array}{array of dimensions k x 3 x n, where k=number of coordinates and n=sample size.}

\item{align}{logical: if TRUE, the data will be aligned first}

\item{use.lm}{integer vector: specifies the indices of the points that are to be used in the constrained model}

\item{deselect}{logical: if TRUE, \code{use.lm} references the missing coordinates instead of the present ones.}

\item{sigma}{estimate of error variance (sensible is a value estimating coordinate error in terms of observer error)}

\item{exVar}{numeric value with \code{0 < exVar <= 1} specifying the PCs to be included by their cumulative explained Variance}

\item{scale}{logical: allow scaling in Procrustes fitting}

\item{representer}{a triangular mesh, where the vertices correspond to the coordinates in \code{array}, leave NULL for pointclouds.}

\item{model}{object of class \code{pPCA}}
}
\value{
returns a probabilistic PCA model as S4 class "pPCA" (see \code{\link{pPCA-class}}).
\code{UpdateModel} is used to modify existing models by changing sigma and exVar.
}
\description{
calculate or modify a probablistic PCA based on 3D-coordinates
}
\examples{
require(Morpho)
data(boneData)
model <- pPCA(boneLM[,,])
## change parameters without recomputing Procrustes fit
model1 <- UpdateModel(model, sigma=1, exVar=0.8)
}
\references{
\enc{Lüthi}{Luethi} M, Albrecht T, Vetter T. 2009. Probabilistic modeling and visualization of the flexibility in morphable models. In: Mathematics of Surfaces XIII. Springer. p 251-264
}

