% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPPCA.r
\docType{methods}
\name{Set-pPCA-class}
\alias{AddModelInfoParams,pPCA-method}
\alias{Set-pPCA-class}
\alias{SetMeanVector<-}
\alias{SetMeanVector<-,pPCA-method}
\alias{SetModelDataInfo<-,pPCA-method}
\alias{SetModelInfoParams<-,pPCA-method}
\alias{SetNoiseVariance<-}
\alias{SetNoiseVariance<-,pPCA-method}
\alias{SetPCA<-}
\alias{SetPCA<-,pPCA-method}
\alias{SetPCBasisMatrix<-}
\alias{SetPCBasisMatrix<-,pPCA-method}
\alias{SetPCsdev<-}
\alias{SetPCsdev<-,pPCA-method}
\alias{SetRawdata<-}
\alias{SetRawdata<-,pPCA-method}
\alias{SetRepresenter<-}
\alias{SetRepresenter<-,pPCA-method}
\alias{SetScale<-}
\alias{SetScale<-,pPCA-method}
\alias{SetScores<-}
\alias{SetScores<-,pPCA-method}
\alias{SetVariance<-}
\alias{SetVariance<-,pPCA-method}
\title{Low level methods to set pPCA class content}
\usage{
SetNoiseVariance(x) <- value

\S4method{SetNoiseVariance}{pPCA}(x) <- value

SetRawdata(x) <- value

\S4method{SetRawdata}{pPCA}(x) <- value

SetPCBasisMatrix(x) <- value

\S4method{SetPCBasisMatrix}{pPCA}(x) <- value

SetPCsdev(x) <- value

\S4method{SetPCsdev}{pPCA}(x) <- value

SetMeanVector(x) <- value

\S4method{SetMeanVector}{pPCA}(x) <- value

SetScores(x) <- value

\S4method{SetScores}{pPCA}(x) <- value

SetRepresenter(x) <- value

\S4method{SetRepresenter}{pPCA}(x) <- value

SetScale(x) <- value

\S4method{SetScale}{pPCA}(x) <- value

SetPCA(x) <- value

\S4method{SetPCA}{pPCA}(x) <- value

SetVariance(x) <- value

\S4method{SetVariance}{pPCA}(x) <- value

\S4method{AddModelInfoParams}{pPCA}(x, value)

\S4method{SetModelInfoParams}{pPCA}(x) <- value

\S4method{SetModelDataInfo}{pPCA}(x) <- value
}
\arguments{
\item{x}{of class "pPCA"}

\item{value}{set the specific value}
}
\value{
returns an updated pPCA object
}
\description{
Low level methods to set pPCA class content - not to be invoked directly
}
\keyword{internals}

