% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setPPCA.r
\docType{methods}
\name{Set-pPCA-class}
\alias{AddModelInfoParams,pPCA-method}
\alias{Set-pPCA-class}
\alias{SetMeanVector<-}
\alias{SetNoiseVariance<-}
\alias{SetPCA<-}
\alias{SetPCBasisMatrix<-}
\alias{SetPCsdev<-}
\alias{SetRawdata<-}
\alias{SetScale<-}
\alias{SetScores<-}
\alias{SetVariance<-}
\title{Low level methods to set pPCA class content}
\usage{
SetNoiseVariance(x) <- value

\S4method{SetNoiseVariance}{pPCA}(x) <- value

SetRawdata(x) <- value

\S4method{SetRawdata}{pPCA}(x) <- value

SetPCBasisMatrix(x) <- value

\S4method{SetPCBasisMatrix}{pPCA}(x) <- value

SetPCsdev(x) <- value

\S4method{SetPCsdev}{pPCA}(x) <- value

SetMeanVector(x) <- value

\S4method{SetMeanVector}{pPCA}(x) <- value

SetScores(x) <- value

\S4method{SetScores}{pPCA}(x) <- value

SetScale(x) <- value

\S4method{SetScale}{pPCA}(x) <- value

SetPCA(x) <- value

\S4method{SetPCA}{pPCA}(x) <- value

SetVariance(x) <- value

\S4method{SetVariance}{pPCA}(x) <- value

\S4method{AddModelInfoParams}{pPCA}(x, value)
}
\arguments{
\item{value}{set the specific value}

\item{model}{of class "pPCA"}
}
\value{
returns an updated pPCA object
}
\description{
Low level methods to set pPCA class content - not to be invoked directly
}
\keyword{internals}

