% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{Set-pPCA-class}
\alias{Set-pPCA-class}
\alias{SetMeanVector<-}
\alias{SetNoiseVariance<-}
\alias{SetPCA<-}
\alias{SetPCBasisMatrix<-}
\alias{SetPCsdev<-}
\alias{SetRawdata<-}
\alias{SetScale<-}
\alias{SetScores<-}
\alias{SetVariance<-}
\title{Low level methods to set pPCA class content}
\usage{
SetNoiseVariance(x) <- value

SetRawdata(x) <- value

SetPCBasisMatrix(x) <- value

SetPCsdev(x) <- value

SetMeanVector(x) <- value

SetScores(x) <- value

SetScale(x) <- value

SetPCA(x) <- value

SetVariance(x) <- value
}
\arguments{
\item{model}{of class "pPCA"}

\item{value}{set the specific value}
}
\value{
returns an updated pPCA object
}
\description{
Low level methods to set pPCA class content - not to be invoked directly
}
\keyword{internals}

