% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictSample.r, R/PredictSamplePointer.r
\docType{methods}
\name{PredictSample}
\alias{PredictSample}
\alias{PredictSample,pPCA,matrix-method}
\alias{PredictSample,pPCA,mesh3d-method}
\alias{PredictSample,pPCA,missing-method}
\alias{PredictSample,pPCA_pointer,matrix-method}
\alias{PredictSample,pPCA_pointer,mesh3d-method}
\alias{PredictSample,pPCA_pointer,missing-method}
\title{predict or restrict a mesh or matrix based on a statistical model}
\usage{
PredictSample(model, dataset, representer = TRUE, ...)

\S4method{PredictSample}{pPCA,matrix}(model, dataset, representer = TRUE,
  origSpace = TRUE, lmDataset = NULL, lmModel = NULL, sdmax = NULL,
  mahaprob = c("none", "chisq", "dist"), align = TRUE, addNoise = FALSE,
  posteriorMean = FALSE, ptValueNoise = 1, ...)

\S4method{PredictSample}{pPCA,mesh3d}(model, dataset, representer = TRUE,
  origSpace = TRUE, lmDataset = NULL, lmModel = NULL, sdmax = NULL,
  mahaprob = c("none", "chisq", "dist"), align = TRUE, addNoise = FALSE,
  posteriorMean = FALSE, ptValueNoise = 1, ...)

\S4method{PredictSample}{pPCA,missing}(model, dataset, representer = TRUE,
  origSpace = TRUE, lmDataset = NULL, lmModel = NULL, sdmax = NULL,
  mahaprob = c("none", "chisq", "dist"), align = TRUE, addNoise = FALSE,
  posteriorMean = FALSE, ptValueNoise = 1, ...)

\S4method{PredictSample}{pPCA_pointer,matrix}(model, dataset,
  representer = TRUE, origSpace = TRUE, lmDataset = NULL,
  lmModel = NULL, sdmax = NULL, mahaprob = c("none", "chisq", "dist"),
  align = TRUE, addNoise = FALSE, posteriorMean = FALSE,
  ptValueNoise = 1, ...)

\S4method{PredictSample}{pPCA_pointer,mesh3d}(model, dataset,
  representer = TRUE, origSpace = TRUE, lmDataset = NULL,
  lmModel = NULL, sdmax = NULL, mahaprob = c("none", "chisq", "dist"),
  align = TRUE, addNoise = FALSE, posteriorMean = FALSE,
  ptValueNoise = 1, ...)

\S4method{PredictSample}{pPCA_pointer,missing}(model, dataset,
  representer = TRUE, origSpace = TRUE, lmDataset = NULL,
  lmModel = NULL, sdmax = NULL, mahaprob = c("none", "chisq", "dist"),
  align = TRUE, addNoise = FALSE, posteriorMean = FALSE,
  ptValueNoise = 1, ...)
}
\arguments{
\item{model}{model of class \code{pPCA}}

\item{dataset}{a matrix or a mesh3d. If \code{dataset} is missing, \code{lmDataset} and \code{lmModel} need to be provided and the posterior mean will be returned.}

\item{representer}{if TRUE and the model contains a representer mesh, a surface mesh will be returned, coordinate matrix otherwise.}

\item{...}{currently not in use.}

\item{origSpace}{logical: rotate the estimation back into the original coordinate system.}

\item{lmDataset}{optional: landmarks on the dataset used for alignment.}

\item{lmModel}{optional: landmarks on the model's mean used for alignment.}

\item{sdmax}{maximum allowed standard deviation (per Principal axis) within the model space. Defines the probabilistic boundaries.}

\item{mahaprob}{character: if != "none", use mahalanobis-distance to determine overall probability (of the shape projected into the model space."chisq" uses the Chi-Square distribution of the squared Mahalanobisdistance, while "dist" restricts the values to be within a multi-dimensional sphere of radius \code{sdmax}. If FALSE the probability will be determined per PC separately.}

\item{align}{if TRUE, the sample will be aligned to the mean.}

\item{addNoise}{re-add noise while reprojecting from latent into shape space.}

\item{posteriorMean}{if TRUE, instead the shape will be the mean of the posterior model using the coordinates defined by \code{lmModel} and \code{lmDataset}.}

\item{ptValueNoise}{numeric: set global noise assumed in the data when \code{posteriorMean = TRUE}}
}
\value{
\code{PredictSample} returns a matrix/mesh3d restricted to the boundaries given by the modelspace.
}
\description{
predict or restrict a mesh or matrix based on a statistical model
}
\seealso{
\code{\link{StatismoModelMembers}}
}
