% Generated by roxygen2 (4.0.1): do not edit by hand
\name{representer2sample}
\alias{representer2sample}
\title{get the representer from a model of class "pPCA"}
\usage{
representer2sample(model)
}
\arguments{
\item{model}{object of class \code{\link{pPCA}}}
}
\value{
an object of class mesh3d or matrix, depending whether a point cloud or a triangular mesh is the model's representer.
}
\description{
get the representer from a model of class "pPCA"
}

