% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alignment.r
\name{rigidAlign}
\alias{rigidAlign}
\title{Fast Procrustes align of coordinates}
\usage{
rigidAlign(array, scale = FALSE, use.lm = NULL, deselect = FALSE)
}
\arguments{
\item{array}{array of coordinates}

\item{scale}{logical: request scaling during alignment}

\item{use.lm}{integer vector: specifies the indices of the points that are to be used in the constrained model}

\item{deselect}{logical: if TRUE, \code{use.lm} references the missing coordinates instead of the present ones.}
}
\value{
a list containing
\item{rotated}{array containing registered coordinates}
\item{mshape}{matrix containing meanshape}
}
\description{
Fast Procrustes align of coordinates
}

