% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Rvtk-package}
\alias{Rvtk-package}
\alias{RvtkStatismo}
\alias{RvtkStatismo-package}
\title{Interface between R and vtk-libraries for mesh operations}
\description{
Aims to provide meshing functionality from vtk for R.
}
\details{
\tabular{ll}{
Package: \tab Rvtk\cr
Type: \tab Package\cr
Version: \tab 0.2.140617\cr
Date: \tab 2014-06-17\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr }
}
\author{
Stefan Schlager

Maintainer: Stefan Schlager <zarquon42@gmail.com>
}
\references{
To be announced
}
\keyword{package}

