% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Rvtk-package}
\alias{Rvtk-package}
\alias{RvtkStatismo}
\alias{RvtkStatismo-package}
\title{Integrates statismo and R using the vtkStandardMeshRepresente}
\description{
Integrates statismo and R using the vtkStandardMeshRepresenter. Statismo shape models will be stored as objects of class "pPCA". (this is work in progress).
}
\details{
\tabular{ll}{
Package: \tab RvtkStatismo\cr
Type: \tab Package\cr
Version: \tab 0.2.140625\cr
Date: \tab 2014-06-25\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr }
}
\author{
Stefan Schlager

Maintainer: Stefan Schlager <zarquon42@gmail.com>
}
\references{
To be announced
}
\keyword{package}

