% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvtkStatismo-package.R
\name{deprecated}
\alias{ComputeCoefficientsForDataset}
\alias{ComputeLogProbabilityOfDataset}
\alias{ComputeMahalanobisDistanceForDataset}
\alias{ComputeProbabilityOfDataset}
\alias{deprecated}
\title{deprecated functions of RvtkStatismo}
\usage{
ComputeProbabilityOfDataset(...)

ComputeLogProbabilityOfDataset(...)

ComputeCoefficientsForDataset(...)

ComputeMahalanobisDistanceForDataset(...)
}
\description{
document deprecated functions
}
\keyword{internal}

