% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvtkStatismo-package.R
\docType{package}
\name{RvtkStatismo-package}
\alias{RvtkStatismo}
\alias{RvtkStatismo-package}
\title{Integrating  statismo and R using the vtkStandardMeshRepresente}
\description{
Integrating statismo and R using the vtkStandardMeshRepresenter.
Statismo shape models will be stored as objects of class "pPCA". (this is work in progress).
}
\details{
\tabular{ll}{
Package: \tab RvtkStatismo\cr
Type: \tab Package\cr
Version: \tab 0.3.160309\cr
Date: \tab 2016-03-09\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr }
}
\author{
Stefan Schlager

Maintainer: Stefan Schlager <zarquon42@gmail.com>
}
\references{
To be announced
}
\keyword{package}

