% Generated by roxygen2 (4.0.1): do not edit by hand
\name{statismoBuildModel}
\alias{statismoBuildModel}
\title{generate a statistical model using an array of superimposed landmarks and save it}
\usage{
statismoBuildModel(array, representer, sigma = 0)
}
\arguments{
\item{array}{array of aligned 3D-coordinates (e.g. the vertices of meshes)}

\item{representer}{matrix or triangular mesh of class "mesh3d" with vertices corresponding to rows in the array.}

\item{sigma}{noise in the data}
}
\description{
generate a statistical model using an array of superimposed landmarks and save it
}

