% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statismoBuildModel.r
\name{statismoBuildModel}
\alias{statismoBuildModel}
\title{generate a statistical model using an array of superimposed landmarks or a list of meshes}
\usage{
statismoBuildModel(x, representer, sigma = 0, scale = FALSE)
}
\arguments{
\item{x}{array of aligned 3D-coordinates or a list of aligned registered meshes.}

\item{representer}{matrix or triangular mesh of class "mesh3d" with vertices corresponding to rows in the array.}

\item{sigma}{noise in the data}

\item{scale}{logical: set to TRUE, if scaling was involved in the registration.}
}
\value{
an object of class pPCA (\code{\link{pPCA-class}})
}
\description{
generate a statistical model using an array of superimposed landmarks
}
\examples{
require(Morpho)
data(boneData)
align <- rigidAlign(boneLM)$rotated
mymod <- statismoBuildModel(align,representer=align[,,1],sigma=2,scale=TRUE)
## save it
statismoSaveModel(mymod,"mymod.h5")
}
\seealso{
\code{\link{pPCA}, \link{pPCA-class}, \link{rigidAlign}, \link{meshalign}}
}
\keyword{StatisticalModel<representer>}

