% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statismoBuildModel.r
\name{statismoBuildModel}
\alias{statismoBuildModel}
\title{generate a statistical model using an array of superimposed landmarks or a list of meshes}
\usage{
statismoBuildModel(
  x,
  representer,
  sigma = 0,
  scale = FALSE,
  computeScores = TRUE,
  SelfAdjointEigenSolver = FALSE,
  pointer = FALSE
)
}
\arguments{
\item{x}{array of aligned 3D-coordinates or a list of aligned registered meshes.}

\item{representer}{matrix or triangular mesh of class "mesh3d" with vertices corresponding to rows in the array.}

\item{sigma}{noise in the data}

\item{scale}{logical: set to TRUE, if scaling was involved in the registration.}

\item{computeScores}{logical: if TRUE PCscores are computed}

\item{SelfAdjointEigenSolver}{logical: if TRUE SelfAdjointEigenSolver is used during model building. Only use this if sample size exceeds amount of variables (faster but less accurate) - can lead to errors otherwise.}

\item{pointer}{if TRUE an object of class pPCA_pointer is returned}
}
\value{
an object of class pPCA (\code{\link{pPCA-class}})
}
\description{
generate a statistical model using an array of superimposed landmarks
}
\examples{
require(Morpho)
data(boneData)
align <- rigidAlign(boneLM)$rotated
mymod <- statismoBuildModel(align,representer=align[,,1],sigma=2,scale=TRUE)
## save it
statismoSaveModel(mymod,"mymod.h5")
}
\seealso{
\code{\link{pPCA}, \link{pPCA-class}, \link{rigidAlign}, \link{meshalign}}
}
\keyword{StatisticalModel<representer>}
