% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{statismoConstrainModelSafe}
\alias{statismoConstrainModelSafe}
\title{calculate a posterior model but only use likely correspondences}
\usage{
statismoConstrainModelSafe(model, sample, pt, ptValueNoise, sdmax = 5)

\S4method{statismoConstrainModelSafe}{pPCA,matrix,numeric}(model, sample, pt,
  ptValueNoise, sdmax = 5)
}
\arguments{
\item{model}{object of class pPCA}

\item{sample}{matrix containing coordinates to constrain model to}

\item{pt}{either a k x 3 matrix with each row containing points on the model's domain corresponding to the row in \code{sample} or an integer vector specifying the coordinates of the sample's mean corresponding to \code{sample}}

\item{ptValueNoise}{numeric: specify noise on constraints.}

\item{sdmax}{a measure in standard deviations to allow the likelihood of the correspondeces between sample and model. (using chi-square distribution)}
}
\value{
a constrained model
}
\description{
calculate a posterior model but only use likely correspondences
}

