% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statismoGaussProcess.r
\name{statismoGPmodel}
\alias{statismoGPmodel}
\title{expands a models variability by adding a Gaussian kernel function}
\usage{
statismoGPmodel(model, kernel = GaussianKernel(50, 10), ncomp = 10,
  nystroem = 500, pointer = FALSE)
}
\arguments{
\item{model}{shape model of class \code{\link{pPCA}}}

\item{kernel}{an object of class matrixKernel}

\item{ncomp}{integer: number of PCs to approximate}

\item{nystroem}{number of samples to compute Nystroem approximation of eigenvectors}

\item{pointer}{if TRUE an object of class pPCA_pointer is returned}
}
\value{
returns a shape model of class \code{\link{pPCA}}
}
\description{
expands a models variability by adding a Gaussian kernel function to the empiric covariance matrix and builds a low-rank approximation of the resulting PCA
}
\examples{
### this is a silly example with only 10 landmarks
require(Morpho)
data(boneData)
align <- rigidAlign(boneLM)$rotated
mod <- statismoBuildModel(align)
##extend flexibility using two Gaussian kernels
## Get the empirical kernel first
isoKernel <- IsoKernel(DrawMean(mod),scale=0.001)
kernel1 <- GaussianKernel(10,1)
kernel2 <- GaussianKernel(1,1)
combinedKernel <- SumKernels(kernel1,kernel2)
combinedKernelNoEmp <- SumKernels(combinedKernel,isoKernel)
combinedKernel <- SumKernels(combinedKernelNoEmp,StatisticalModelKernel())
GPmod <- statismoGPmodel(mod,kernel=combinedKernel)
##extend flexibility using two Gaussian kernels but ignoring empiric covariance.
GPmodNoEmp <- statismoGPmodel(mod,kernel=combinedKernelNoEmp)
PC1orig <- DrawSample(mod,2)# get shape in 2sd of first PC of originial model
PC1 <- DrawSample(GPmod,2)# get shape in 2sd of first PC of the extended model
PC1NoEmp <- DrawSample(GPmodNoEmp,2)# get shape in 2sd of first PC
##visualize the differences from the mean (green spheres)
\dontrun{
deformGrid3d(PC1,DrawMean(GPmod),ngrid=0)##
##only deviates in 5 landmarks from the mean (dark blue)
deformGrid3d(PC1NoEmp,DrawMean(GPmod),ngrid=0,col1=4,add=TRUE)
deformGrid3d(PC1orig,DrawMean(GPmod),ngrid=0,col1=5,add=TRUE)
}
}
\seealso{
\code{\link{pPCA}, \link{pPCA-class}}
}
\keyword{StatisticalModel<representer>}

