% Generated by roxygen2 (4.0.2): do not edit by hand
\name{statismoGPmodel}
\alias{statismoGPmodel}
\title{expands a models variability by adding a Gaussian kernel function}
\usage{
statismoGPmodel(model, useEmpiric = TRUE, kernel = list(c(100, 70)),
  ncomp = 10, nystroem = 500, combine = c("sum", "product"),
  combineEmp = "sum")
}
\arguments{
\item{model}{shape model of class \code{\link{pPCA}}}

\item{useEmpiric}{logical: if TRUE, the empiric covariance kernel will be added to the Gaussian ones.}

\item{kernel}{a list containing two valued vectors containing with the first entry specifiying the bandwidth and the second the scaling of the Gaussian kernels.}

\item{ncomp}{integer: number of PCs to approximate}

\item{nystroem}{number of samples to compute Nystroem approximation of eigenvectors}

\item{combine}{character: determining how to combine the kernels: "sum" or "product" are supported.}

\item{combineEmp}{character: determining how the newly created kernel is combined to the one from the model: "sum" or "product" are supported.}
}
\value{
returns a shape model of class \code{\link{pPCA}}
}
\description{
expands a models variability by adding a Gaussian kernel function to the empiric covariance matrix and builds a low-rank approximation of the resulting PCA
}
\examples{
### this is a silly example with only 10 landmarks
require(Morpho)
data(boneData)
align <- rigidAlign(boneLM)$rotated
mod <- statismoBuildModel(align)
GPmod <- statismoGPmodel(mod,kernel=list(c(10,1),c(1,1)))##extend flexibility using two Gaussian kernels
GPmodNoEmp <- statismoGPmodel(mod,kernel=list(c(10,1),c(1,1)),useEmpiric = FALSE)##extend flexibility using two Gaussian kernels but ignoring empiric covariance.
PC1orig <- DrawSample(mod,2)# get shape in 2sd of first PC of originial model
PC1 <- DrawSample(GPmod,2)# get shape in 2sd of first PC of the extended model
PC1NoEmp <- DrawSample(GPmodNoEmp,2)# get shape in 2sd of first PC
##visualize the differences from the mean (green spheres)
deformGrid3d(PC1,DrawMean(GPmod),ngrid=0)##
deformGrid3d(PC1NoEmp,DrawMean(GPmod),ngrid=0,col1=4,add=TRUE)##only deviates in 5 landmarks from the mean (dark blue)
deformGrid3d(PC1orig,DrawMean(GPmod),ngrid=0,col1=5,add=TRUE)
}
\seealso{
\code{\link{pPCA}, \link{pPCA-class}}
}
\keyword{StatisticalModel<representer>}

