% Generated by roxygen2 (4.0.1): do not edit by hand
\name{statismoSaveModel}
\alias{statismoLoadModel}
\alias{statismoSaveModel}
\title{save and load a statistical model of class pPCA to statismo hdf5 format}
\usage{
statismoSaveModel(model, modelname = dataname)

statismoLoadModel(modelname, scale = TRUE)
}
\arguments{
\item{model}{object of class pPCA}

\item{modelname}{filename to read/save}
}
\value{
statismoLoadModel returns an object of class "pPCA" while statismoSaveModel saves an object of class pPCA to disk in the statismo file format.
}
\description{
save and load a statistical model of class pPCA to statismo hdf5 format
}

