% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statismoIO.r
\name{statismoLoadModel/statismoSaveModel}
\alias{statismoLoadModel/statismoSaveModel}
\alias{statismoSaveModel}
\alias{statismoLoadModel}
\title{save and load a statistical model of class pPCA to statismo hdf5 format}
\usage{
statismoSaveModel(model, modelname = dataname)

statismoLoadModel(modelname, pointer = FALSE, scale = FALSE)
}
\arguments{
\item{model}{object of class \code{\link{pPCA}}}

\item{modelname}{filename to read/save}

\item{pointer}{if TRUE an object of class pPCA_pointer is returned}

\item{scale}{specify if scaling was involved in model generation}
}
\value{
statismoLoadModel returns an object of class \code{\link{pPCA}} while statismoSaveModel saves an object of class \code{\link{pPCA}} to disk in the statismo file format.
}
\description{
save and load a statistical model of class pPCA to statismo hdf5 format
}
\seealso{
\code{\link{pPCA}}
}
\keyword{StatisticalModel<representer>}
