% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r, R/ModelMembers.r,
%   R/ModelMembersPointer.r
\docType{methods}
\name{StatismoMatrices}
\alias{StatismoMatrices}
\alias{GetPCABasisMatrix}
\alias{GetOrthonormalPCABasisMatrix}
\alias{GetCovarianceAtPoint}
\alias{GetCovarianceMatrix}
\alias{GetJacobian}
\alias{GetProjectionMatrix}
\alias{GetPCABasisMatrix,pPCA-method}
\alias{GetOrthonormalPCABasisMatrix,pPCA-method}
\alias{GetCovarianceAtPoint,pPCA,numeric,numeric-method}
\alias{GetCovarianceMatrix,pPCA-method}
\alias{GetJacobian,pPCA,numeric-method}
\alias{GetProjectionMatrix,pPCA-method}
\alias{GetPCABasisMatrix,pPCA_pointer-method}
\alias{GetOrthonormalPCABasisMatrix,pPCA_pointer-method}
\alias{GetCovarianceAtPoint,pPCA_pointer,numeric,numeric-method}
\alias{GetCovarianceMatrix,pPCA_pointer-method}
\alias{GetJacobian,pPCA_pointer,numeric-method}
\alias{GetProjectionMatrix,pPCA_pointer-method}
\title{Get Matrices from StatisticalModel class}
\usage{
GetPCABasisMatrix(model)

GetOrthonormalPCABasisMatrix(model)

GetCovarianceAtPoint(model, pt1, pt2)

GetCovarianceMatrix(model)

GetJacobian(model, pt)

GetProjectionMatrix(model)

\S4method{GetPCABasisMatrix}{pPCA}(model)

\S4method{GetOrthonormalPCABasisMatrix}{pPCA}(model)

\S4method{GetCovarianceAtPoint}{pPCA,numeric,numeric}(model, pt1, pt2)

\S4method{GetCovarianceMatrix}{pPCA}(model)

\S4method{GetJacobian}{pPCA,numeric}(model, pt)

\S4method{GetProjectionMatrix}{pPCA}(model)

\S4method{GetPCABasisMatrix}{pPCA_pointer}(model)

\S4method{GetOrthonormalPCABasisMatrix}{pPCA_pointer}(model)

\S4method{GetCovarianceAtPoint}{pPCA_pointer,numeric,numeric}(model, pt1,
  pt2)

\S4method{GetCovarianceMatrix}{pPCA_pointer}(model)

\S4method{GetJacobian}{pPCA_pointer,numeric}(model, pt)

\S4method{GetProjectionMatrix}{pPCA_pointer}(model)
}
\arguments{
\item{model}{model of class "pPCA"}

\item{pt1}{either an integer pointing to the index of the domain or a numeric vector of length 3 specifying a point on the domain of the model}

\item{pt2}{either an integer pointing to the index of the domain or a numeric vector of length 3 specifying a point on the domain of the model}

\item{pt}{either an integer pointing to the index of the domain or a numeric vector of length 3 specifying a point on the domain of the model}
}
\value{
\item{GetPCABasisMatrix}{returns the (scaled) Basis of the latent space}
\item{GetOrthonormalPCABasisMatrix}{ returns the orthonormal Basis of the latent space}
\item{GetCovarianceMatrix}{returns the covariance matrix - can be huge!!!}
 \item{GetCovarianceAtPoint}{returns the 3 x 3 covariance matrix for \code{pt1} and \code{pt2}}
\item{GetJacobian}{ returns the 3 x 3 Jacobian matrix at \code{pt}}
\item{GetProjectionMatrix}{ returns matrix to project a sample vector into the latent space (this is not a member function but might prove useful anyway)}
}
\description{
Get Matrices from StatisticalModel class - such as projection matrices, covariance matrices or Jacobian
}
\details{
see \url{http://statismo.github.io/docs/api/v0.10/html/classstatismo_1_1StatisticalModel.html} for details.
}
