% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{StatismoModelMembers}
\alias{ComputeCoefficientsForDataset}
\alias{ComputeCoefficientsForPointValues}
\alias{ComputeLogProbabilityOfDataset}
\alias{ComputeProbabilityOfDataset}
\alias{DrawMean}
\alias{DrawSample}
\alias{GetCovarianceAtPoint}
\alias{GetCovarianceMatrix}
\alias{GetDomainPoints}
\alias{GetDomainSize}
\alias{GetJacobian}
\alias{GetMeanVector}
\alias{GetNoiseVariance}
\alias{GetOrthonormalPCABasisMatrix}
\alias{GetPCAVarianceVector}
\alias{StatismoModelMembers}
\title{Implementation/Emulation of the statismo StatisticalModel class.}
\usage{
GetPCABasisMatrix(model)

GetOrthonormalPCABasisMatrix(model)

GetNoiseVariance(model)

GetMeanVector(model)

GetPCAVarianceVector(model)

ComputeLogProbabilityOfDataset(model, dataset)

ComputeProbabilityOfDataset(model, dataset)

DrawMean(model)

DrawSample(model, coefficients = NULL, addNoise = FALSE)

ComputeCoefficientsForDataset(model, dataset)

GetDomainPoints(model)

GetDomainSize(model)

GetCovarianceAtPoint(model, pt1, pt2)

GetCovarianceMatrix(model)

GetJacobian(model, pt)

ComputeCoefficientsForPointValues(model, sample, ids)
}
\arguments{
\item{model}{object of class \code{\link{pPCA}}}

\item{dataset}{an (already aligned) mesh or k x 3 matrix containing the datasets coordinates.}

\item{coefficients}{specify coefficients in the latent space to draw a sample}

\item{addNoise}{logical: if TRUE noise as specified in the model will be added to the returned sample}
}
\value{
functions return matrices, (log)-probabilties, coefficients or sample (mesh3d or matrix) for specific dataset
}
\description{
Implementation/Emulation of the statismo StatisticalModel class.
}
\details{
see \url{http://statismo.github.io/statismo/classdoc/html/classstatismo_1_1StatisticalModel.html} for details.
}
\keyword{StatisticalModel<representer>}

