% Generated by roxygen2 (4.0.1): do not edit by hand
\name{StatismoModelMembers}
\alias{ComputeCoefficientsForDataset}
\alias{ComputeCoefficientsForPointValues}
\alias{DrawMeanAtPoint}
\alias{DrawSample}
\alias{DrawSampleAtPoint}
\alias{DrawSampleVector}
\alias{EvaluateSampleAtPoint}
\alias{GetCovarianceAtPoint}
\alias{GetDomainPoints}
\alias{GetDomainSize}
\alias{StatismoModelMembers}
\title{Implementation/Emulation of the statismo StatisticalModel class.}
\usage{
DrawMean(model)

DrawMeanAtPoint(model, pt)

DrawSample(model, coefficients = NULL, addNoise = FALSE)

DrawSampleVector(model, coefficients, addNoise = FALSE)

DrawSampleAtPoint(model, coefficients, pt, addNoise = FALSE)

ComputeCoefficientsForDataset(model, dataset)

ComputeCoefficientsForPointValues(model, sample, pt, ptNoise = 0)

GetDomainPoints(model)

GetDomainSize(model)

EvaluateSampleAtPoint(model, sample, pt)

GetCovarianceAtPoint(model, pt1, pt2)
}
\arguments{
\item{model}{object of class \code{\link{pPCA}}}

\item{dataset}{an (already aligned) mesh or k x 3 matrix containing the datasets coordinates.}

\item{coefficients}{specify coefficients in the latent space to draw a sample}

\item{addNoise}{logical: if TRUE noise as specified in the model will be added to the returned sample}

\item{ptNoise}{specify the noise estimated in the points.}
}
\value{
\item{DrawMean}{Get the mean (either a matrix or a mesh3d)}
\item{GetMeanVector}{Get the mean vector}
\item{DrawMeanAtPoint}{Get a specific point of the  mean (numeric vector)}
\item{DrawSample}{Draw a sample from the model (either a matrix or a mesh3d)}
\item{DrawMeanAtPoint}{Get a specific point of the  mean (numeric vector)}
\item{DrawSampleAtPoint}{Draw a sample of a specific point from the model (numeric vector)  }
\item{ComputeCoefficientsForDataset}{Computes the coefficients of the latent variables}
\item{ComputeCoefficientsForPointValues}{Returns the coefficients of the latent variables for the given values provided in two k x 3 matrices or two vectors of length 3, or one matrix/vector and a vector containing the indices on the domain  corresponding to these points}
\item{GetDomainPoints}{a matrix containing the points of the model's domain}
\item{GetDomainSize}{get the size of the model's domain}
\item{EvaluateSampleAtPoint}{Returns the value of the given sample at the point specified (either as point on the domain or as an index)}
}
\description{
Implementation/Emulation of the statismo StatisticalModel class.
}
\details{
see \url{http://statismo.github.io/statismo/classdoc/html/classstatismo_1_1StatisticalModel.html} for details.
}
\keyword{StatisticalModel<representer>}

