% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.r, R/ModelMembers.r
\docType{methods}
\name{StatismoModelMembers}
\alias{ComputeCoefficientsForDataset}
\alias{ComputeCoefficientsForDataset,pPCA-method}
\alias{ComputeCoefficientsForPointValues}
\alias{ComputeCoefficientsForPointValues,pPCA,matrix,matrix,numeric-method}
\alias{ComputeCoefficientsForPointValues,pPCA,matrix,numeric,numeric-method}
\alias{ComputeCoefficientsForPointValues,pPCA,numeric,numeric,numeric-method}
\alias{DrawMean}
\alias{DrawMean,pPCA-method}
\alias{DrawMeanAtPoint}
\alias{DrawMeanAtPoint,pPCA,numeric-method}
\alias{DrawSample}
\alias{DrawSample,pPCA-method}
\alias{DrawSampleAtPoint}
\alias{DrawSampleAtPoint,pPCA,numeric,numeric-method}
\alias{DrawSampleVector}
\alias{DrawSampleVector,pPCA-method}
\alias{EvaluateSampleAtPoint}
\alias{EvaluateSampleAtPoint,pPCA,matrix,numeric-method}
\alias{EvaluateSampleAtPoint,pPCA,mesh3d,numeric-method}
\alias{GetDomainPoints}
\alias{GetDomainPoints,pPCA-method}
\alias{GetDomainSize}
\alias{GetDomainSize,pPCA-method}
\alias{GetModelInfo}
\alias{GetModelInfo,pPCA-method}
\alias{GetPCScores}
\alias{GetPCScores,pPCA-method}
\alias{StatismoModelMembers}
\title{Implementation/Emulation of the statismo StatisticalModel class.}
\usage{
DrawMean(model)

DrawMeanAtPoint(model, pt)

DrawSample(model, coefficients = NULL, addNoise = FALSE)

DrawSampleVector(model, coefficients, addNoise = FALSE)

DrawSampleAtPoint(model, coefficients, pt, addNoise = FALSE)

ComputeCoefficientsForDataset(model, dataset)

ComputeCoefficientsForPointValues(model, sample, pt, ptNoise = 0)

GetDomainPoints(model)

GetDomainSize(model)

EvaluateSampleAtPoint(model, sample, pt)

GetModelInfo(model)

GetPCScores(model, scaled = TRUE)

\S4method{DrawMean}{pPCA}(model)

\S4method{DrawMeanAtPoint}{pPCA,numeric}(model, pt)

\S4method{DrawSample}{pPCA}(model, coefficients = NULL, addNoise = FALSE)

\S4method{DrawSampleVector}{pPCA}(model, coefficients, addNoise = FALSE)

\S4method{DrawSampleAtPoint}{pPCA,numeric,numeric}(model, coefficients, pt,
  addNoise = FALSE)

\S4method{ComputeCoefficientsForDataset}{pPCA}(model, dataset)

\S4method{GetDomainPoints}{pPCA}(model)

\S4method{GetDomainSize}{pPCA}(model)


  \S4method{ComputeCoefficientsForPointValues}{pPCA,matrix,numeric,numeric}(model,
  sample, pt, ptNoise = 0)

\S4method{ComputeCoefficientsForPointValues}{pPCA,matrix,matrix,numeric}(model,
  sample, pt, ptNoise = 0)


  \S4method{ComputeCoefficientsForPointValues}{pPCA,numeric,numeric,numeric}(model,
  sample, pt, ptNoise = 0)

\S4method{EvaluateSampleAtPoint}{pPCA,matrix,numeric}(model, sample, pt)

\S4method{EvaluateSampleAtPoint}{pPCA,mesh3d,numeric}(model, sample, pt)

\S4method{GetModelInfo}{pPCA}(model)

\S4method{GetPCScores}{pPCA}(model, scaled = TRUE)
}
\arguments{
\item{model}{object of class \code{\link{pPCA}}}

\item{pt}{either an integer pointing to a coordinate or a 3D-vector containing the coordinates of the domain point of interest. For \code{ComputeCoefficientsForPointValues}, this can also specify a matrix of coordinates on the domain.}

\item{coefficients}{specify coefficients in the latent space to draw a sample}

\item{addNoise}{logical: if TRUE noise as specified in the model will be added to the returned sample}

\item{dataset}{an (already aligned) mesh or k x 3 matrix containing the datasets coordinates.}

\item{sample}{depending on the function a matrix, a numeric vector or a mesh3d (see methods below)}

\item{ptNoise}{specify the noise estimated in the points.}

\item{scaled}{logical: if TRUE, the scores are scaled by their standard deviation.}
}
\value{
\item{DrawMean}{Get the mean (either a matrix or a mesh3d)}
\item{GetMeanVector}{Get the mean vector}
\item{DrawMeanAtPoint}{Get a specific point of the  mean (numeric vector)}
\item{DrawSample}{Draw a sample from the model (either a matrix or a mesh3d)}
\item{DrawMeanAtPoint}{Get a specific point of the  mean (numeric vector)}
\item{DrawSampleAtPoint}{Draw a sample of a specific point from the model (numeric vector)  }
\item{ComputeCoefficientsForDataset}{Computes the coefficients of the latent variables}
\item{ComputeCoefficientsForPointValues}{Returns the coefficients of the latent variables for the given values provided in two k x 3 matrices or two vectors of length 3, or one matrix/vector and a vector containing the indices on the domain  corresponding to these points}
\item{GetDomainPoints}{a matrix containing the points of the model's domain}
\item{GetDomainSize}{get the size of the model's domain}
\item{EvaluateSampleAtPoint}{Returns the value of the given sample at the point specified (either as point on the domain or as an index)}
\item{GetPCScores}{get model's PC-scores, scaled or unscaled to unit variance, depending on the choice of \code{scaled}}
}
\description{
Implementation/Emulation of the statismo StatisticalModel class.

#
}
\details{
see \url{http://statismo.github.io/docs/api/v0.10/html/classstatismo_1_1StatisticalModel.html} for details.
}
\keyword{StatisticalModel<representer>}

