% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statismoModelFromRepresenter.r
\name{statismoModelFromRepresenter}
\alias{statismoModelFromRepresenter}
\title{generate model from a representer using gaussian kernels}
\usage{
statismoModelFromRepresenter(representer, kernel = GaussianKernel(100,
  50), ncomp = 10, nystroem = 500, pointer = FALSE)
}
\arguments{
\item{representer}{mesh3d or matrix used as representer}

\item{kernel}{an object of class matrixKernel}

\item{ncomp}{integer: number of PCs to approximate}

\item{nystroem}{number of samples to compute Nystroem approximation of eigenvectors}

\item{pointer}{if TRUE an object of class pPCA_pointer is returned}
}
\value{
returns a shape model of class \code{\link{pPCA}}
}
\description{
generate model from a representer using gaussian kernels
}
\examples{
require(Rvcg)
data(humface)
kernel1 <- GaussianKernel(80,50)
kernel2 <- GaussianKernel(20,20)
combinedKernel <- SumKernels(kernel1,kernel2)
hummodel <- statismoModelFromRepresenter(humface,combinedKernel)
\dontrun{
require(rgl)
for (i in 1:5) wire3d(DrawSample(hummodel),col=i)
}
}
