% Generated by roxygen2 (4.0.1): do not edit by hand
\name{statismoModelFromRepresenter}
\alias{statismoModelFromRepresenter}
\title{generate model from a representer using gaussian kernels}
\usage{
statismoModelFromRepresenter(representer, kernel = list(c(100, 70)),
  ncomp = 10, nystroem = 500, combine = "sum")
}
\arguments{
\item{representer}{mesh3d or matrix used as representer}

\item{kernel}{a list containing two valued vectors containing with the first entry specifiying the bandwidth and the second the scaling of the Gaussian kernels.}

\item{ncomp}{integer: number of PCs to approximate}

\item{nystroem}{number of samples to compute Nystroem approximation of eigenvectors}

\item{combine}{character determining how to combine the kernels: "sum" or "product" are supported.}
}
\value{
returns a shape model of class \code{\link{pPCA}}
}
\description{
generate model from a representer using gaussian kernels
}
\examples{
require(Rvcg)
data(humface)
hummodel <- statismoModelFromRepresenter(humface)
require(rgl)
for (i in 1:5) wire3d(DrawSample(hummodel),col=i)
}

