% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{StatismoParameters}
\alias{GetMeanVector}
\alias{GetMeanVector,pPCA-method}
\alias{GetNoiseVariance}
\alias{GetNoiseVariance,pPCA-method}
\alias{GetPCAVarianceVector}
\alias{GetPCAVarianceVector,pPCA-method}
\alias{StatismoParameters}
\title{Get model parameters}
\usage{
GetNoiseVariance(model)

GetMeanVector(model)

GetPCAVarianceVector(model)

\S4method{GetNoiseVariance}{pPCA}(model)

\S4method{GetMeanVector}{pPCA}(model)

\S4method{GetPCAVarianceVector}{pPCA}(model)
}
\arguments{
\item{model}{model of class "pPCA"}
}
\value{
\item{GetNoiseVariance}{returns the estimated noise in the model}
\item{GetPCAVarianceVector}{returns the variance in the model}
\item{GetMeanVector}{returns the model's mean vector}
}
\description{
Get model parameters such as variance or noise variance
}
\details{
see \url{http://statismo.github.io/statismo/classdoc/html/classstatismo_1_1StatisticalModel.html} for details.
}

