% Generated by roxygen2 (4.0.1): do not edit by hand
\name{statismoReducedVariance}
\alias{statismoReducedVariance}
\title{Reduce an existing statistical shape model}
\usage{
statismoReducedVariance(model, exVar = 1, npc = 0, scores = TRUE)
}
\arguments{
\item{model}{}

\item{exVar}{restricts model by explained variance - with \code{0 < exVar < 1}}

\item{npc}{number of PCs retained in the model (overrides \code{exVar})}

\item{scores}{logical: request recomputation of PC-scores}
}
\description{
Reduce an existing statistical shape model either to first n PCs or by explained Variance
}
\examples{
require(Morpho)
data(boneData)
align <- rigidAlign(boneLM)$rotated
mymod <- statismoBuildModel(align,representer=align[,,1],sigma=2,scale=TRUE)
reducemod <- statismoReducedVariance(mymod,0.5)
}

