% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReducedVarianceModel.r
\docType{methods}
\name{statismoReducedVariance}
\alias{statismoReducedVariance}
\alias{statismoReducedVariance,pPCA-method}
\title{Reduce an existing statistical shape model}
\usage{
statismoReducedVariance(model, exVar = 1, npc = 0, scores = TRUE)

\S4method{statismoReducedVariance}{pPCA}(model, exVar = 1, npc = 0)
}
\arguments{
\item{model}{statistical shape model of class 'pPCA'}

\item{exVar}{restricts model by explained variance - with \code{0 < exVar < 1}}

\item{npc}{number of PCs retained in the model (overrides \code{exVar})}

\item{scores}{logical: if TRUE, the scores for the reduced model will be returned.}
}
\description{
Reduce an existing statistical shape model either to first n PCs or by explained Variance
}
\examples{
require(Morpho)
data(boneData)
align <- rigidAlign(boneLM)$rotated
mymod <- statismoBuildModel(align,representer=align[,,1],sigma=2,scale=TRUE)
reducemod <- statismoReducedVariance(mymod,0.5)
}

