% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r, R/ModelMembers.r
\docType{methods}
\name{StatismoSample}
\alias{ComputeLogProbabilityOfDataset}
\alias{ComputeLogProbabilityOfDataset,pPCA-method}
\alias{ComputeMahalanobisDistanceForDataset}
\alias{ComputeMahalanobisDistanceForDataset,pPCA-method}
\alias{ComputeProbabilityOfDataset}
\alias{ComputeProbabilityOfDataset,pPCA-method}
\alias{StatismoSample}
\title{Retrieve information about a sample from the model}
\usage{
ComputeLogProbabilityOfDataset(model, dataset)

ComputeProbabilityOfDataset(model, dataset)

ComputeMahalanobisDistanceForDataset(model, dataset)

\S4method{ComputeLogProbabilityOfDataset}{pPCA}(model, dataset)

\S4method{ComputeProbabilityOfDataset}{pPCA}(model, dataset)

\S4method{ComputeMahalanobisDistanceForDataset}{pPCA}(model, dataset)
}
\arguments{
\item{model}{model of class "pPCA"}

\item{dataset}{a matrix or mesh3d aligned to the model's mean}
}
\value{
\item{ComputeLogProbabilityOfDataset}{returns the log-probability density for the sample}
\item{ComputeProbabilityOfDataset}{returns the probability density for the sample}
}
\description{
Retrieve information about a sample from the model
}
\seealso{
\code{\link{getDataLikelihood}}
}

