% Generated by roxygen2 (4.0.1): do not edit by hand
\name{statismoSaveModel}
\alias{statismoSaveModel}
\title{save a statistical model of class pPCA to statismo hdf5 format}
\usage{
statismoSaveModel(model, modelname)
}
\arguments{
\item{model}{object of class pPCA}

\item{modelname}{filename to save to}
}
\description{
save a statistical model of class pPCA to statismo hdf5 format
}

