% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkGeodesicPath.r
\name{vtkGeodesicPath}
\alias{vtkGeodesicPath}
\title{Compute Geodesic path between two vertices}
\usage{
vtkGeodesicPath(x, start, end)
}
\arguments{
\item{x}{triangular mesh of class mesh3d}

\item{start}{index of vertex to start}

\item{end}{index of vertex to end}
}
\value{
list containing
\item{distance}{geodesic distance (cummulative distance between all vertices along the way)}
\item{index}{vector containing vertex indices from start to end}
}
\description{
Compute Geodesic path between two vertices
}

