% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtkImageBlend.r
\name{vtkImageBlend}
\alias{vtkImageBlend}
\title{Blend two Images using vtkImageBlend}
\usage{
vtkImageBlend(image1, image2, outname = "out.mha", opacity1 = 0.5,
  opacity2 = 0.5)
}
\arguments{
\item{image1}{character: path to first image}

\item{image2}{character: path to second image}

\item{outname}{character: image to save merged image to.}

\item{opacity1}{opacity of image1. Accepts values between 0 and 1.}

\item{opacity2}{opacity of image2. Accepts values between 0 and 1.}
}
\description{
Blend two Images using vtkImageBlend
}

