% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtkImageBlend.r
\name{vtkImageResize}
\alias{vtkImageResize}
\title{Resample an image and write to file.}
\usage{
vtkImageResize(image, outname, spacing = c(1, 1, 1), interpolate = 2)
}
\description{
Resample an image and write to file.
}

