% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkMeshVolume.r
\name{vtkMeshInfo}
\alias{vtkMeshInfo}
\title{get Infos about a triangular mesh}
\usage{
vtkMeshInfo(mesh)
}
\arguments{
\item{mesh}{triangular mesh}
}
\value{
\item{volume}{enclosed volume}
\item{projVol}{enclosed projected volume}
\item{surfaceArea}{surface area}
}
\description{
get Infos about a triangular mesh
}
\details{
see documentation of vtkMassProperties for details
}
