% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtkstuff.r
\name{vtkMeshWrite}
\alias{vtkMeshWrite}
\title{exports a triangular mesh of class mesh3d to a vtp file}
\usage{
vtkMeshWrite(mesh, filename = dataname, type = c("vtk", "vtp"),
  ascii = FALSE)
}
\arguments{
\item{mesh}{mesh of class mesh3d}

\item{filename}{character}

\item{type}{character: file extension. Can be "vtk", "vtp".}

\item{ascii}{if TRUE, vtk will be written as binary}
}
\description{
exports a triangular mesh of class mesh3d to a vtp file
}

